requirements.txt
asgiref==3.5.2
autopep8==1.6.0
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==4.0.5
django-environ==0.9.0
django-heroku==0.3.1
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
psycopg2==2.9.3
pycodestyle==2.8.0
pytz==2022.1
requests==2.28.0
slack-sdk==3.17.1
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.9
whitenoise==6.2.0
